auto refresh for community posts
